relocated addnewlinkss api to modifylinks path
