+++ serversideapp/Controller/UserLinksEditsController/UserLinksModifyController.js
@@ -2,11 +2,14 @@ import express from 'express';
 import {
 	callAddLinkService,
 	deleteUserLinksService,
+	addNewLinkExistingUserService,
 } from '../../Service/UserLinksModifyServices/UserLinksModifyServices.js';
 import {
 	CheckIfValidDeleteLinkRequest,
 	getLinksForUserValidatorService,
 } from '../../Service/ServiceValidator/CheckIfValidRequest.js';
+import { validateAddLinkExistingUserRequest } from '../../Service/ServiceValidator/UserLinksServiceValidator.js';
+import { UserNotFoundErrorMessage } from '../../General/ErrorMessages/APIErrorrMessages.js';
 const router = express.Router();
 
 router.route('/addLink').post((req, res) => {
@@ -29,6 +32,27 @@ router.route('/deleteLinkExistingUser').patch((req, res) => {
 	}
 
 	deleteUserLinksService(res, userIdFromReq, userNameFromReq, listOfLinksIdToRemove);
+});
+
+router.route('/addLinkExistingUser').patch((req, res) => {
+	const userNameFromReq = req.body.userName;
+	const userDataLinksSchemaLocal = req.body.userDataObject;
+
+	const userNameFilter = { userName: userNameFromReq };
+	const userNotFoundError = UserNotFoundErrorMessage(userNameFromReq);
+	// console.log('here before');
+	if (!validateAddLinkExistingUserRequest(res, userNameFromReq, userDataLinksSchemaLocal)) {
+		console.log('here in if');
+		return;
+	}
+	console.log('here in addLinkExistingUser');
+
+	addNewLinkExistingUserService(
+		res,
+		userNameFilter,
+		userDataLinksSchemaLocal,
+		userNotFoundError
+	);
 });
 
 export { router as userLinksEditRouter };
